timebank visual view funcitonality

+++ activity/src/main/java/com/kairos/persistence/model/wta/templates/template_types/AverageScheduledTimeWTATemplate.java
@@ -141,7 +141,7 @@ public class AverageScheduledTimeWTATemplate extends WTABaseRuleTemplate {
                         totalMin +=dateTimeInterval.overlap(shift.getDateTimeInterval()).getMinutes();
                     }
                 }
-                boolean isValid = isValid(minMaxSetting, limitAndCounter[0], totalMin/(60*dateTimeInterval.getDays()));
+                boolean isValid = isValid(minMaxSetting, limitAndCounter[0], totalMin/(60*(int)dateTimeInterval.getDays()));
                 if (!isValid) {
                     if(limitAndCounter[1]!=null) {
                         int counterValue =  limitAndCounter[1] - 1;
